Support for teaching teams

+++ nexus/app/controllers/deadline_extension_controller.rb
@@ -1,12 +1,11 @@
 class DeadlineExtensionController < ApplicationController
   include ApplicationHelper
 
-  before_action :authenticate_admin!
-
   def create
     @deadline_extension = DeadlineExtension.new(deadline_extension_params)
-
     if @deadline_extension.save
+      return unless authenticate_can_administrate!(@deadline_extension.assignment.course)
+
       flash[:success] = 'Deadline extension created!'
       redirect_to assignment_deadline_extensions_path(id: @deadline_extension.assignment.id)
       @deadline_extension.assignment.log("Deadline Extension created for #{@deadline_extension.user.name} until #{strftime_uk @deadline_extension.extendeddeadline}")
@@ -17,11 +16,14 @@ class DeadlineExtensionController < ApplicationController
 
   def edit
     @deadline_extension = DeadlineExtension.find(params[:id])
+    authenticate_can_administrate!(@deadline_extension.assignment.course) if @deadline_extension
   end
 
   def update
     @deadline_extension = DeadlineExtension.find(params[:id])
     if @deadline_extension.update_attributes(deadline_extension_params)
+      return unless authenticate_can_administrate!(@deadline_extension.assignment.course)
+
       flash[:success] = 'Deadline extension updated'
       redirect_to assignment_deadline_extensions_path(id: @deadline_extension.assignment.id)
       @deadline_extension.assignment.log("Deadline Extension updated for #{@deadline_extension.user.name} until #{strftime_uk @deadline_extension.extendeddeadline}")
@@ -32,20 +34,24 @@ class DeadlineExtensionController < ApplicationController
 
   def destroy
     @deadline_extension = DeadlineExtension.find(params[:id])
+    if @deadline_extension
+      return unless authenticate_can_administrate!(@deadline_extension.assignment.course)
 
-    @assignment = @deadline_extension.assignment
+      @assignment = @deadline_extension.assignment
 
-    if @deadline_extension.destroy
-      flash[:success] = 'Deadline extension revoked'
-      @assignment.log("Deadline Extension revoked for #{@deadline_extension.user.name}")
-    else
-      flash[:error] = 'Error revoking deadline extension'
+      if @deadline_extension.destroy
+        flash[:success] = 'Deadline extension revoked'
+        @assignment.log("Deadline Extension revoked for #{@deadline_extension.user.name}")
+      else
+        flash[:error] = 'Error revoking deadline extension'
+      end
+      redirect_to assignment_deadline_extensions_path(id: @assignment.id)
     end
-    redirect_to assignment_deadline_extensions_path(id: @assignment.id)
   end
 
   def new
     @deadline_extension = DeadlineExtension.new(assignment_id: params[:aid])
+    authenticate_can_administrate!(@deadline_extension.assignment.course) if @deadline_extension
   end
 
   private
